some naming refactor
