+++ cli/plasmo/src/features/project-creator/from-existing-manifest.ts
