+++ cli/plasmo/src/features/project-creator/get-raw-name.ts
