+++ cli/plasmo/src/features/project-creator/install-dependencies.ts
